Set wasm stack pointer prior to running any code.
With the wasm backend the initial value of the SP global is set by
wasm-ld, but then emscripten adds static data using makeStaticAlloc
which causes the stack and heap to shift upwards in the address range.
However were never updatig the value in the SP global in wasm binary.

With this change we set the SP on startup.  Its hard to write a test
for this since the effect of this bug is basically that you effective
stack range is a few tens of bytes less.

If did verify by had the value on startup is otherwise the one that
wasm-ld assigned.